Initial weapon breakage processor.

+++ src/main/java/org/labcabrera/rolemaster/core/model/item/Item.java
@@ -47,7 +47,7 @@ public class Item {
 
 	private Range<BigDecimal> commonWeight;
 
-	private Range<BigDecimal> commonLength;
+	private Range<Integer> commonLength;
 
 	private List<ItemFeature> commonFeatures;
